+++ app/src/main/java/com/aa/draftt/viewmodels/HomeViewModel.kt
@@ -2,7 +2,7 @@ package com.aa.draftt.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.aa.draftt.auth.repositories.FirebaseAuthRepository
+import com.aa.draftt.repositories.implementations.FirebaseAuthRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
